+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Improved accessibility of back to top button by moving focus to top.
 - Improved accessibility of pagination buttons.
 - Improved accessibility of external links in link block.
+- Improved accessibility of external links in highlight block.
 - Fixed back icon in pagination.
 
 ## 1.0.1